Moved c_test_file into its own file
