+++ src/test-c-gen/Makefile
@@ -1,4 +1,4 @@
-SRC = c_test_case_generator.cpp c_test_source_factory.cpp expr2cleanc.cpp function_parameter_builder.cpp function_return_builder.cpp
+SRC = c_test_case_generator.cpp c_test_source_factory.cpp expr2cleanc.cpp function_parameter_builder.cpp function_return_builder.cpp c_test_file.cpp
 
 INCLUDES= -I ..
